+++ docs/source/conf.py
@@ -24,9 +24,9 @@ copyright = '2020 MERIDIAN'
 author = 'Fabio Frazao, Oliver Kirsebom'
 
 # The short X.Y version
-version = '2.3'
+version = '2.4'
 # The full version, including alpha/beta/rc tags
-release = '2.3.1'
+release = '2.4.0'
 
 
 # -- General configuration ---------------------------------------------------